+++ data-transfer-object/src/main/java/org/grnet/status/dtos/tenant/TenantResponseDto.java
@@ -6,6 +6,7 @@ import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.grnet.status.dtos.tenant.metadata.TenantMetadata;
 import org.grnet.status.dtos.tenant.status.TenantStatusDto;
+import org.grnet.status.enums.TenantGroupStatus;
 
 import java.util.List;
 
@@ -59,5 +60,14 @@ public class TenantResponseDto {
     @JsonProperty("status")
     @Valid
     public TenantStatusDto status;
+
+
+    @Schema(
+            description = "Indicates whether the tenant group exists in the IAM system",
+            implementation = String.class,
+            example = "EXISTS"
+    )
+    @JsonProperty("group-status")
+    public TenantGroupStatus groupStatus;
 
 }