+++ enum/src/main/java/org/grnet/status/enums/TenantGroupStatus.java
@@ -1,0 +1,7 @@
+package org.grnet.status.enums;
+
+public enum TenantGroupStatus {
+    EXISTS,
+    NOT_FOUND,
+    UNKNOWN
+}